+++ GUI-dev/src/app/pages/pages-routing.module.ts
@@ -4,6 +4,7 @@ import { NgModule } from '@angular/core';
 import { PagesComponent } from './pages.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { HomeComponent } from './home/home.component';
+import { ProcessStoreSearchComponent } from './processstore-search/processstore-search.component';
 import {AuthGuard} from '../auth-guard.service';
 
 
@@ -14,6 +15,11 @@ const routes: Routes = [{
     {
       path: 'dashboard',
       component: DashboardComponent,
+      canActivate: [AuthGuard],
+    },
+    {
+      path: 'processstore-search',
+      component: ProcessStoreSearchComponent,
       canActivate: [AuthGuard],
     },
     {