+++ GUI-dev/src/app/pages/pages.module.ts
@@ -5,6 +5,7 @@ import { DashboardModule } from './dashboard/dashboard.module';
 import { PagesRoutingModule } from './pages-routing.module';
 import { ThemeModule } from '../@theme/theme.module';
 import {HomeModule} from './home/home.module';
+import { ProcessStoreSearchModule } from './processstore-search/processstore-search.module';
 
 const PAGES_COMPONENTS = [
   PagesComponent,
@@ -16,6 +17,7 @@ const PAGES_COMPONENTS = [
     ThemeModule,
     DashboardModule,
     HomeModule,
+    ProcessStoreSearchModule,
   ],
   declarations: [
     ...PAGES_COMPONENTS,